Refine the interface for C
